BE/Fix : 드리미 리뷰 테이블 score을 integer로 변경

+++ backend/sql/sym-boorm-ddl.sql
@@ -315,7 +315,7 @@ CREATE TABLE `DREAMI_REQUEST_DENIED_DETAILS` (
 
 CREATE TABLE `DREAMI_REVIEW` (
                                  `review_id`    binary(16)    NOT NULL,
-                                 `score`        tinyint       NOT NULL,
+                                 `score`        integer       NOT NULL,
                                  `content`      varchar(200)  NULL,
                                  `created_dtm`  timestamp     NOT NULL  DEFAULT CURRENT_TIMESTAMP,
                                  `updated_dtm`  timestamp     NULL,
